# openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Agent Gateway API
  version: 0.7.0
  description: Unified sessions, messages, interactions, cancellation, and SSE events for replaceable coding agents.
servers:
  - url: http://127.0.0.1:3000
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
  /health:
    get:
      security: []
      summary: Liveness probe
      responses:
        "200":
          description: Gateway process is alive
          content:
            application/json:
              schema:
                type: object
                required: [status, engine, version]
                properties:
                  status: { type: string, const: ok }
                  engine: { type: string }
                  version: { type: string }
  /ready:
    get:
      security: []
      summary: Readiness probe
      responses:
        "200": { description: Gateway and repository are ready }
        "503": { description: Gateway is shutting down or repository is unavailable }
  /openapi.yaml:
    get:
      security: []
      summary: Get the HTTP API contract
      responses:
        "200": { description: OpenAPI document }
  /asyncapi.yaml:
    get:
      security: []
      summary: Get the SSE event contract
      responses:
        "200": { description: AsyncAPI document }
  /metrics:
    get:
      summary: Prometheus metrics
      responses:
        "200":
          description: Prometheus text exposition
          content:
            text/plain:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/engines:
    get:
      summary: Get active and available engines
      responses:
        "200":
          description: Engine metadata
          content:
            application/json:
              schema:
                type: object
                required: [gatewayVersion, active, capabilities, available, engines]
                properties:
                  gatewayVersion: { type: string }
                  active: { type: string }
                  available: { type: array, items: { type: string } }
                  capabilities: { $ref: "#/components/schemas/AgentCapabilities" }
                  engines:
                    type: array
                    items: { $ref: "#/components/schemas/EngineDescriptor" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/events:
    get:
      summary: Subscribe to the global or session-filtered SSE stream
      parameters:
        - in: query
          name: sessionId
          schema: { type: string }
        - in: query
          name: lastEventId
          description: Replay retained events after this sequence number. The Last-Event-ID header is also supported.
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: Server-Sent Events stream. Each data field contains a GatewayEvent JSON object.
          content:
            text/event-stream:
              schema: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/sessions:
    get:
      summary: List sessions across configured engines
      responses:
        "200":
          description: Session list
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions: { type: array, items: { $ref: "#/components/schemas/Session" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Create a session
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                directory:
                  type: string
                  description: Existing project directory, restricted by AGENT_ALLOWED_ROOTS when configured.
                engine:
                  type: string
                  description: Optional engine for this session. Defaults to the engine selected by --engine.
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                type: object
                required: [session]
                properties:
                  session: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /v1/sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      summary: Get a session and its messages
      responses:
        "200":
          description: Session snapshot
          content:
            application/json:
              schema:
                type: object
                required: [session]
                properties:
                  session: { $ref: "#/components/schemas/Session" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      summary: Delete a session and stop an active run
      responses:
        "204": { description: Session deleted }
        "404": { $ref: "#/components/responses/Error" }
  /v1/sessions/{sessionId}/messages:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      summary: Start an Agent generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, minLength: 1 }
      responses:
        "202":
          description: Generation accepted in queued or running state; subscribe to SSE for updates
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunAccepted" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /v1/sessions/{sessionId}/interactions/{requestId}/respond:
    parameters:
      - $ref: "#/components/parameters/SessionId"
      - in: path
        name: requestId
        required: true
        schema: { type: string }
    post:
      summary: Answer an Agent question or permission request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/QuestionResponse"
                - $ref: "#/components/schemas/PermissionResponse"
      responses:
        "202": { description: Interaction response accepted }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /v1/sessions/{sessionId}/stop:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      summary: Stop the active generation
      responses:
        "202":
          description: Stop accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunAccepted" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /v1/sessions/{sessionId}/runs:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      summary: List runs for a session
      responses:
        "200":
          description: Runs in reverse creation order
          content:
            application/json:
              schema:
                type: object
                required: [runs]
                properties:
                  runs: { type: array, items: { $ref: "#/components/schemas/Run" } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/sessions/{sessionId}/interactions:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      summary: List persisted interactions for a session
      responses:
        "200":
          description: Interactions in reverse creation order
          content:
            application/json:
              schema:
                type: object
                required: [interactions]
                properties:
                  interactions: { type: array, items: { $ref: "#/components/schemas/Interaction" } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/runs/{runId}:
    parameters:
      - in: path
        name: runId
        required: true
        schema: { type: string }
    get:
      summary: Get a canonical run
      responses:
        "200":
          description: Run snapshot
          content:
            application/json:
              schema:
                type: object
                required: [run]
                properties:
                  run: { $ref: "#/components/schemas/Run" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/runs/{runId}/interactions:
    parameters:
      - in: path
        name: runId
        required: true
        schema: { type: string }
    get:
      summary: List persisted interactions for a run
      responses:
        "200":
          description: Interactions in reverse creation order
          content:
            application/json:
              schema:
                type: object
                required: [interactions]
                properties:
                  interactions: { type: array, items: { $ref: "#/components/schemas/Interaction" } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/interactions/{interactionId}:
    parameters:
      - in: path
        name: interactionId
        required: true
        schema: { type: string }
    get:
      summary: Get a persisted interaction
      responses:
        "200":
          description: Interaction snapshot
          content:
            application/json:
              schema:
                type: object
                required: [interaction]
                properties:
                  interaction: { $ref: "#/components/schemas/Interaction" }
        "404": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    SessionId:
      in: path
      name: sessionId
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Gateway error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: API key missing or invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    AgentCapabilities:
      type: object
      required: [protocol, protocolVersion, nativeSessions, questions, permissions, cancellation]
      properties:
        protocol: { type: string }
        protocolVersion: { type: string }
        nativeSessions: { type: boolean }
        questions: { type: boolean }
        permissions: { type: boolean }
        cancellation: { type: boolean }
    EngineDescriptor:
      type: object
      required: [name, displayName, capabilities]
      properties:
        name: { type: string }
        displayName: { type: string }
        capabilities: { $ref: "#/components/schemas/AgentCapabilities" }
    ContentPart:
      oneOf:
        - type: object
          required: [type, text]
          properties:
            type: { type: string, const: text }
            text: { type: string }
        - type: object
          required: [type, value]
          properties:
            type: { type: string, const: json }
            value: {}
        - type: object
          required: [type, name]
          properties:
            type: { type: string, const: file }
            name: { type: string }
            path: { type: string }
            mimeType: { type: string }
        - type: object
          required: [type, artifactId]
          properties:
            type: { type: string, const: artifact }
            artifactId: { type: string }
    Message:
      type: object
      required: [id, role, content, status, createdAt]
      properties:
        id: { type: string }
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
        parts: { type: array, items: { $ref: "#/components/schemas/ContentPart" } }
        status: { type: string, enum: [completed, stopped] }
        createdAt: { type: string, format: date-time }
    Session:
      type: object
      required: [id, engine, directory, status, messages, createdAt, updatedAt]
      properties:
        id: { type: string }
        engine: { type: string }
        directory: { type: string }
        workspace:
          type: object
          required: [type, directory]
          properties:
            type: { type: string, const: local }
            directory: { type: string }
        status: { type: string, enum: [idle, busy] }
        messages: { type: array, items: { $ref: "#/components/schemas/Message" } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Run:
      type: object
      required: [id, sessionId, engine, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        sessionId: { type: string }
        engine: { type: string }
        status:
          type: string
          enum: [queued, running, input_required, canceling, completed, failed, canceled]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        outputMessageId: { type: string }
        stopReason: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
    Interaction:
      type: object
      required: [id, sessionId, runId, type, status, data, createdAt, updatedAt]
      properties:
        id: { type: string }
        sessionId: { type: string }
        runId: { type: string }
        type: { type: string, enum: [question, permission] }
        status: { type: string, enum: [pending, resolved, canceled] }
        data: {}
        response:
          oneOf:
            - $ref: "#/components/schemas/QuestionResponse"
            - $ref: "#/components/schemas/PermissionResponse"
        resolvedBy: { type: string, enum: [client, policy] }
        cancelReason: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time }
    RunAccepted:
      type: object
      required: [runId]
      properties:
        runId: { type: string }
    QuestionResponse:
      type: object
      properties:
        answer: { type: string }
        answers: { type: object, additionalProperties: true }
    PermissionResponse:
      type: object
      properties:
        decision: { type: string, enum: [allow, deny] }
        optionId: { type: string }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
            requestId: { type: string }