# src/index.ts
export { createApp, type AppOptions } from "./app.js";
export { GatewayService } from "./gateway-service.js";
export { loadGatewayConfig, type GatewayConfig, type PermissionPolicy } from "./config.js";
export {
  MemorySessionRepository,
  SqliteSessionRepository,
  createSessionRepository,
  type SessionRepository,
} from "./session-store.js";
export {
  MemoryRunRepository,
  SqliteRunRepository,
  createRunRepository,
  type RunRepository,
} from "./run-store.js";
export {
  MemoryInteractionRepository,
  SqliteInteractionRepository,
  createInteractionRepository,
  type InteractionRepository,
} from "./interaction-store.js";
export {
  MemoryEventRepository,
  SqliteEventRepository,
  createEventRepository,
  type EventRepository,
} from "./event-store.js";
export {
  EngineCatalog,
  type EngineCatalogLike,
  type EngineDescriptor,
} from "./engines/catalog.js";
export {
  createEngine,
  availableEngines,
  registerEngineProtocol,
  type EngineDefinition,
  type EngineName,
  type EngineProtocol,
  type EngineFactory,
} from "./engines/registry.js";
export type {
  AgentCapabilities,
  AgentEngine,
  AgentRunContext,
  AgentSessionContext,
} from "./engines/types.js";
export type {
  ContentPart,
  GatewayEvent,
  GatewayEventType,
  Message,
  Interaction,
  InteractionStatus,
  InteractionType,
  Run,
  RunStatus,
  Session,
  WorkspaceRef,
} from "./types.js";
export { WorkspaceResolver } from "./workspace.js";
export {
  NoopTransactionCoordinator,
  SqliteDatabase,
  type TransactionCoordinator,
} from "./sqlite-database.js";